requirements.txt
botbuilder-core
aiohttp
